now in fixed-point for f a s t
i'm using an oscilloscope for perf measurements (don't hit me), because
it shows the timing between consecutive dmas. i've gotten it down from
22ms (for 100 leds) to just a bit over 8ms \o/

i totally expected this thing to have an fpu, i'm sorry

+++ Core/Inc/main.h
@@ -99,7 +99,7 @@ void render_frame(void *data, FrameBuffer *fb, uint32_t time);
 
 void clear_framebuffer(FrameBuffer *fb);
 int framebuffer_push_rgb(FrameBuffer *fb, uint8_t red, uint8_t green, uint8_t blue);
-int framebuffer_push_hsv(FrameBuffer *fb, double h, double s, double l);
+int framebuffer_push_hsv(FrameBuffer *fb, uint8_t h, uint8_t s, uint8_t l);
 /* USER CODE END EFP */
 
 /* Private defines -----------------------------------------------------------*/
